Refactor DefaultJvmSoftwareComponent to move towards ownership model
This moves us closer to a goal of the Component being a "dumb" container that doesn't reach out to the project to configure things.

In this change, a Feature (the main Feature) is created independently of the component by the plugin and passed to the component's constructor as an argument.

The component no longer configures publishing on itself (the plugin does this).